refactor(studio): extract AnonymizerBuilderForm from the route
Move the form UI, tab state, and submit handling out of the route into
AnonymizerBuilderForm, rendered under FormProvider. The component can now
call useDefaultRoleModels directly, so the RoleModelDefaults null-render
shim that only existed to reach form context is gone.

The route's separate useModelsListProviders call was identical to the one
inside useAnonymizerModels, so the submit button's loading state now comes
from useDefaultRoleModels instead.

Also import FC with the inline type modifier in the touched files.
